Refactor the install.py and upgrade.py scripts
Primarily, move the bulk of them into installation/.
